requirements.txt
faiss-cpu
sentence-transformers
transformers
numpy